requirements.txt
numpy
pytorch
matplotlib
torchvision
GPUtil
